refactor(engine): relocate barriers + restart_policy to src/tuners/engine/
Introduce the tuner-agnostic execution engine package src/tuners/engine/ and
move the two strategy-independent primitives into it:

- GenerationBarrier (B1-B17 lockstep synchronisation) from src/tuner/core/
- should_restart / restart policy from src/tuner/benchmark/

Both depend only on src.utils, are consumed by every tuner (PBT, BO, LHS) and
the per-worker orchestrator, and belong below the individual strategies in the
dependency graph. The new engine/__init__.py re-exports the public surface.

- git mv both files (history preserved)
- repoint all importers (core, benchmark, main, lhs_design, bo_baseline, tests)
- src/tuner/benchmark/__init__.py keeps re-exporting should_restart for the
  transitional import surface

Part of ADR-006 tuners/ unification. Relocate-first; no behavior change.
