+++ src/tuners/engine/__init__.py
@@ -1,0 +1,23 @@
+"""
+Tuner Execution Engine
+======================
+
+Tuner-agnostic execution primitives shared across all tuning strategies:
+
+- Lockstep generation barriers (B1–B17 synchronisation)
+- TuningMode-driven restart policy
+
+These modules are consumed by every tuner (PBT, BO, LHS) and by the
+per-worker orchestration pipeline. They depend only on ``src.utils`` — never
+on a specific tuning strategy — so they sit below the individual tuners in the
+dependency graph.
+"""
+
+from src.tuners.engine.barriers import GenerationBarrier, BARRIER_NAMES
+from src.tuners.engine.restart_policy import should_restart
+
+__all__ = [
+    "GenerationBarrier",
+    "BARRIER_NAMES",
+    "should_restart",
+]
