+++ src/tuners/lhs_design.py
@@ -35,7 +35,7 @@ from typing import Any, Dict, List, Optional, Tuple
 from concurrent.futures import ThreadPoolExecutor, as_completed
 
 from src.config.data_root import resolve_data_root
-from src.tuner.core.barriers import GenerationBarrier
+from src.tuners.engine.barriers import GenerationBarrier
 from src.tuner.core.worker import Worker
 from src.tuners.base import BaseTuner
 from src.tuners.utils.exceptions import (
