+++ tests/unit/bo/test_cotenant.py
@@ -11,7 +11,7 @@ from unittest.mock import MagicMock
 
 from src.config.database import DatabaseConfig
 from src.knobs import get_knob_space
-from src.tuner.core.barriers import GenerationBarrier
+from src.tuners.engine.barriers import GenerationBarrier
 from src.scripts.bo_baseline.cotenant import CoTenantLoadController
